fix(real-data): remove redundant dtype cast in RegionClientWindowDataset; report verified smoke results

+++ real_data_experiments/common/region_tensor_dataset.py
@@ -27,8 +27,8 @@ class RegionClientWindowDataset(Dataset):
         if input_length <= 0 or horizon <= 0:
             raise ValueError("input_length and horizon must be positive.")
 
-        self.tensor = tensor.detach().to(dtype=torch.float32)
-        self._own_tensor = False  # shared reference, no clone to avoid O(n_clones) memory/time
+        self.tensor = tensor.detach()
+        self._own_tensor = False  # shared reference; dtype conversion done once at bundle load
         self.region_ids = [int(region_id) for region_id in region_ids]
         self.input_length = int(input_length)
         self.horizon = int(horizon)
